separate main drivers for local and global registration; option to read the steps from the bigstream.config
