+++ Dockerfile
@@ -1,5 +1,5 @@
 # Create final image
-FROM ghcr.io/janeliascicomp/dask:2023.10.1-py11-ol9
+FROM ghcr.io/janeliascicomp/dask:2024.4.1-py11-ol9
 ARG TARGETPLATFORM
 
 RUN dnf install -y \
